+++ CHANGELOG.md
@@ -4,6 +4,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.1] - 2021-04-30
+### Changed
+- When setting the `colors` property the gradient now updates to match.
+
+## [3.1.0] - 2021-01-02
+### Added
+- Support for pausing and resuming animation.
+- Support for auto-resuming animation on application becoming active.
+
+## [3.0.0] - 2020-09-22
+### Added
+- Support for Xcode 12.
+
+### Changed
+- Dropped support for iOS 8.
+
 ## [2.1.0] - 2019-11-12
 ### Added
 - Support for [Swift Package Manager](https://swift.org/package-manager/).